beta 1.2.2 它总算能跑了

+++ robot.cpp
@@ -1,6 +1,6 @@
 #include "robot.h"
 
-robot::robot(int _id, float x, float y) {
+robot::robot(int _id, double x, double y) {
     this->ID = _id;
     this->positionX = x;
     this->positionY = y;
